+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Unreleased
 
+## [0.19.2]
+- Adjusted `Accordion`
+- Adjust `RichChat` and `LLMChat` to use "outlined" `Accordion`
+
 ## [0.19.1]
 - Removed Pixi.js integration demo
 
@@ -268,6 +272,7 @@ responsive logic uses the selected `Surface` element to handle persistent margin
 ## [v0.2.1]
 ### Other
 - vibe coded
+[v0.19.2]: https://github.com/off-court-creations/valet/releases/tag/v0.19.2
 [v0.19.1]: https://github.com/off-court-creations/valet/releases/tag/v0.19.1
 [v0.19.0]: https://github.com/off-court-creations/valet/releases/tag/v0.19.0
 [v0.18.3]: https://github.com/off-court-creations/valet/releases/tag/v0.18.3